==> code/01_2_getRefNonRef.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.colors as pltc
import geopandas
import dataretrieval.nwis as nwis
import contextily as cx
from matplotlib.ticker import FormatStrFormatter
import matplotlib as mpl
mpl.rcParams['pdf.fonttype'] = 42

npPNW = pd.read_csv('../data/PNW_DryingRegimeGages.csv')

classDat = pd.read_csv('../data/mean_annual_no_flow_and_climate_metric_means_for_no_flow_sites_082819_with_info.csv')
classDat = classDat[['site', 'CLASS']]
classDat.columns = ['gage', 'CLASS']

NP = pd.merge(npPNW,classDat,on='gage')

NP.to_csv('../data/PNW_DryingRegimeGages.csv')
